feat: align API with LocalStorage, ensure warning is logged only once

+++ src/createMMKV.web.ts
@@ -5,7 +5,19 @@ import { createTextEncoder } from './createTextEncoder';
 const canUseDOM =
   typeof window !== 'undefined' && window.document?.createElement != null;
 
+const hasAccessToLocalStorage = () => {
+  try {
+    // throws ACCESS_DENIED error
+    window.localStorage;
+
+    return true;
+  } catch {
+    return false;
+  }
+};
+
 const KEY_WILDCARD = '\\';
+const inMemoryStorage = new Map<string, string>();
 
 export const createMMKV = (config: MMKVConfiguration): NativeMMKV => {
   if (config.encryptionKey != null) {
@@ -15,6 +27,12 @@ export const createMMKV = (config: MMKVConfiguration): NativeMMKV => {
     throw new Error("MMKV: 'path' is not supported on Web!");
   }
 
+  if (!hasAccessToLocalStorage()) {
+    console.warn(
+      'MMKV: LocalStorage has been disabled. Your experience will be limited to in-memory storage!'
+    );
+  }
+
   const storage = () => {
     if (!canUseDOM) {
       throw new Error(
@@ -22,23 +40,16 @@ export const createMMKV = (config: MMKVConfiguration): NativeMMKV => {
       );
     }
 
-    try {
-      // throws ACCESS_DENIED error
-      window.localStorage;
-    } catch {
-      console.warn(
-        'MMKV: LocalStorage has been disabled. Your experience will be limited to in-memory storage!'
-      );
-
-      const inMemoryStorage = new Map<string, string>();
-
+    if (!hasAccessToLocalStorage()) {
       return {
-        getItem: (key: string) => inMemoryStorage.get(key),
+        getItem: (key: string) => inMemoryStorage.get(key) ?? null,
         setItem: (key: string, value: string) =>
           inMemoryStorage.set(key, value),
         removeItem: (key: string) => inMemoryStorage.delete(key),
         clear: () => inMemoryStorage.clear(),
-      };
+        length: inMemoryStorage.size,
+        key: (index: number) => Object.keys(inMemoryStorage).at(index) ?? null,
+      } as Storage;
     }
 
     const domStorage =
